+++ llvm/include/llvm/SYCLLowerIR/CompileTimePropertiesPass.h
