+++ llvm/include/llvm/SYCLLowerIR/DeviceGlobals.h
