+++ llvm/lib/SYCLLowerIR/CompileTimeProperties.def
