+++ llvm/test/SYCLLowerIR/CompileTimePropertiesPass/kernel-attributes/kernel-pipelined.ll
@@ -1,5 +1,5 @@
 ; Check conversion of sycl-pipelined attribute
-; RUN: sycl-post-link --device-globals --ir-output-only %s -S -o - | FileCheck %s --check-prefix CHECK-IR
+; RUN: opt -passes="compile-time-properties" %s -S -o - | FileCheck %s --check-prefix CHECK-IR
 
 ; CHECK-IR-DAG: @pipelineNegative() #0 {{.*}}!spirv.Decorations [[DEFAULT_PIPELINE:![0-9]+]] {
 ; Function Attrs: convergent norecurse
