+++ llvm/test/SYCLLowerIR/CompileTimePropertiesPass/kernel-attributes/streaming-interface.ll
@@ -1,5 +1,5 @@
 ; Check conversion of sycl-streaming-interface attribute
-; RUN: sycl-post-link --device-globals --ir-output-only %s -S -o - | FileCheck %s --check-prefix CHECK-IR
+; RUN: opt -passes="compile-time-properties" %s -S -o - | FileCheck %s --check-prefix CHECK-IR
 
 ; CHECK-IR-DAG: @pStreaming() #0 {{.*}}!ip_interface [[STREAMING:![0-9]+]] {
 ; Function Attrs: convergent norecurse
