+++ llvm/test/tools/sycl-post-link/emit_program_metadata.ll
@@ -6,10 +6,6 @@
 
 target triple = "spir64-unknown-unknown"
 
-attributes #0 = { "sycl-work-group-size"="4,2,1" }
-attributes #1 = { "sycl-work-group-size"="4,2" }
-attributes #2 = { "sycl-work-group-size"="4" }
-
 !0 = !{i32 1, i32 2, i32 4}
 !1 = !{i32 2, i32 4}
 !2 = !{i32 4}
@@ -19,27 +15,12 @@ define weak_odr spir_kernel void @SpirKernel1(float %arg1) !reqd_work_group_size
   ret void
 }
 
-define weak_odr spir_kernel void @SpirKernel2(float %arg1) #0 {
+define weak_odr spir_kernel void @SpirKernel2(float %arg1) !reqd_work_group_size !1 {
   call void @foo(float %arg1)
   ret void
 }
 
-define weak_odr spir_kernel void @SpirKernel3(float %arg1) !reqd_work_group_size !1 {
-  call void @foo(float %arg1)
-  ret void
-}
-
-define weak_odr spir_kernel void @SpirKernel4(float %arg1) #1 {
-  call void @foo(float %arg1)
-  ret void
-}
-
-define weak_odr spir_kernel void @SpirKernel5(float %arg1) !reqd_work_group_size !2 {
-  call void @foo(float %arg1)
-  ret void
-}
-
-define weak_odr spir_kernel void @SpirKernel6(float %arg1) #2 {
+define weak_odr spir_kernel void @SpirKernel3(float %arg1) !reqd_work_group_size !2 {
   call void @foo(float %arg1)
   ret void
 }
@@ -49,11 +30,8 @@ declare void @foo(float)
 ; CHECK-PROP: [SYCL/program metadata]
 ; // Base64 encoding in the prop file (including 8 bytes length):
 ; CHECK-PROP-NEXT: SpirKernel1@reqd_work_group_size=2|gBAAAAAAAAQAAAAACAAAAQAAAAA
-; CHECK-PROP-NEXT: SpirKernel2@reqd_work_group_size=2|gBAAAAAAAAQAAAAACAAAAQAAAAA
-; CHECK-PROP-NEXT: SpirKernel3@reqd_work_group_size=2|ABAAAAAAAAgAAAAAEAAAAA
-; CHECK-PROP-NEXT: SpirKernel4@reqd_work_group_size=2|ABAAAAAAAAgAAAAAEAAAAA
-; CHECK-PROP-NEXT: SpirKernel5@reqd_work_group_size=2|gAAAAAAAAAABAAAA
-; CHECK-PROP-NEXT: SpirKernel6@reqd_work_group_size=2|gAAAAAAAAAABAAAA
+; CHECK-PROP-NEXT: SpirKernel2@reqd_work_group_size=2|ABAAAAAAAAgAAAAAEAAAAA
+; CHECK-PROP-NEXT: SpirKernel3@reqd_work_group_size=2|gAAAAAAAAAABAAAA
 
 ; CHECK-TABLE: [Code|Properties]
 ; CHECK-TABLE-NEXT: {{.*}}files_0.prop
