+++ llvm/tools/sycl-post-link/sycl-post-link.cpp
@@ -13,8 +13,6 @@
 // - specialization constant intrinsic transformation
 //===----------------------------------------------------------------------===//
 
-#include "CompileTimePropertiesPass.h"
-#include "DeviceGlobals.h"
 #include "ModuleSplitter.h"
 #include "SYCLDeviceLibReqMask.h"
 #include "SYCLDeviceRequirements.h"
@@ -37,6 +35,7 @@
 #include "llvm/IRReader/IRReader.h"
 #include "llvm/Linker/Linker.h"
 #include "llvm/Passes/PassBuilder.h"
+#include "llvm/SYCLLowerIR/DeviceGlobals.h"
 #include "llvm/SYCLLowerIR/ESIMD/LowerESIMD.h"
 #include "llvm/SYCLLowerIR/LowerInvokeSimd.h"
 #include "llvm/SYCLLowerIR/LowerKernelProps.h"
@@ -846,11 +845,6 @@ processInputModule(std::unique_ptr<Module> M) {
       DUMP_ENTRY_POINTS(MDesc2.entries(), MDesc2.Name.c_str(), 3);
       Modified |= processSpecConstants(MDesc2);
 
-      // TODO: detach compile-time properties from device globals.
-      if (DeviceGlobals.getNumOccurrences() > 0) {
-        Modified |=
-            runModulePass<CompileTimePropertiesPass>(MDesc2.getModule());
-      }
       if (!MDesc2.isSYCL() && LowerEsimd) {
         assert(MDesc2.isESIMD() && "NYI");
         // ESIMD lowering also detects large-GRF kernels, so it must happen
